make cli execution better
