+++ sample-notificator-app/src/main/groovy/life/qbic/samplenotificator/NotificatorApp.groovy
@@ -1,28 +1,27 @@
 package life.qbic.samplenotificator
 
 import groovy.util.logging.Log4j2
-import life.qbic.samplenotificator.cli.CommandLineParser
+import life.qbic.business.notification.create.CreateNotificationInput
+import life.qbic.business.notification.create.CreateNotificationOutput
 import life.qbic.samplenotificator.cli.NotificatorCommandLineOptions
 
 /**
- * <b>Entry point to the sample notificator application</b>
- *
- * <p>//TODO</p>
+ * <b>Builds up the app content and starts it</b>
  *
  * @since 1.0.0
  */
 @Log4j2
-class NotificatorApp {
-    public static void main(String[] args) {
-        try {
-            NotificatorCommandLineOptions commandLineParameters =
-                    CommandLineParser.parseAndVerifyCommandLineParameters(args)
+class NotificatorApp implements Runnable{
+    private NotificatorCommandLineOptions commandLineParameters
 
-            DependencyManager dependencyManager = new DependencyManager(commandLineParameters)
-            dependencyManager.run()
-        }catch(Exception exception){
-            log.error "Could not run sample-notificator-cli"
-        }
+    NotificatorApp (NotificatorCommandLineOptions commandLineParameters){
+        this.commandLineParameters = commandLineParameters
     }
 
+    @Override
+    void run() {
+        DependencyManager dependencyManager = new DependencyManager(commandLineParameters)
+        CreateNotificationInput createNotification = dependencyManager.getCreateNotification()
+        createNotification.createNotifications()
+    }
 }
